[check-clean] Add a simple command to list dirty repos.
This functions very similar to `jiri status` but it only cares about
uncommitted changes and will exit non-zero if they exist. This will be
used by Infra to detect if changes in CQ either post-checkout or
post-build leave the source tree in an unclean state. This is not
particularly robust and if a malicious actor wanted to make changes and
run `git commit` the command woult miss it.

Change-Id: Iefbd883cbd0bb91d319be7e424ff60e6fab02d0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/626768
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Nathan Mulcahey <nmulcahey@google.com>