[project] Add use local manifest flag for jiri projects
We want to have an updated jiri projects state without running jiri
update when the manifest is updated. When submodules are enabled, we use
the projects state generated from the new manifest so the
submodule-update tool can figure out what changes need to be made for
submodules.

Bug: b/309855579
Bug: b/301072531
Change-Id: I35509ebd2668886c984fa9860bd9b91fc55bffde
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/945271
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Rahul Bangar <rahulbn@google.com>
Commit-Queue: Yuping Zhai <yupingz@google.com>