[cmdline] Add -timefile flag
Timing information is often quite verbose and clutters up the end of the
Jiri output as viewed in infra, making error messages difficult to find.

Support writing it to a separate file so that it can still be presented
in infra, but separate from the main log.

Bug: b/288448934
Change-Id: Id34773380cbe14d6c9cc468c3763f1270ab8143d
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/963132
Reviewed-by: Drew Fisher <zarvox@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>