[gitutil] Fix compatibility with git<2.31.0
The --path-format flag was only added to git in version 2.31.0 (released
in 2021). Many open-source contributors in particular are on older
versions of git and it's a bad experience to require a super up-to-date
Git in order to check out Fuchsia when it's easy enough to fix the
backwards compatibility.

Bug: 323233422
Change-Id: I430cc686c5368812a2cd2cbcb8e68288b799f969
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/984153
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>