Change default branch to "main" everywhere
commit 97a7c02e832dddab434431f5629e9450da685972 was my first attempt at
changing Jiri to stop uploading to the "master" branch by default. That
was insufficient because Jiri's logic that determines which branch to
push to falls back to the relevant project's `RemoteBranch` attribute,
which still defaulted to "master" (and wasn't overridden in any
manifests).

Changing the default `RemoteBranch` for projects is a much more invasive
change and it was difficult to determine exactly which "master"
references needed to be updated, so I went ahead and replaced all
"master" references in jiri's codebase with "main".

Aside from replacing "master" references with "main", the only code
change is in `project/project.go`, updating Jiri to change the default
`git push` target branch from "master" to "main" by overwriting
.git/config.

I confirmed locally that this fixes `jiri upload` for repos that don't
have a "main" branch, and that `jiri update` continues to work.

Bug: 327615030
Change-Id: I30f3e534b698aff4b98834781e87c29abbcd63fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1038794
Reviewed-by: Danielle Kay <danikay@google.com>
Commit-Queue: Danielle Kay <danikay@google.com>