Remove support for running arbitrary "jiri-*" scripts
Previously, if you ran `jiri foo` and had an executable named `jiri-foo`
on your $PATH, jiri would run that script. This was presumably intended
to mimic Git, which supports the same thing for `git foo` -> `git-foo`,
but it was an undocumented feature and I'm not aware of any people using
it, and can't find any "jiri-*" scripts in Code Search.

This feature may have made sense if someone had grand plans for Jiri to
be a widely-used general-purpose source control layer, since it makes it
easy to extend Jiri with "custom" commands, but I don't think it makes
sense anymore.

If anyone is using this feature, any such commands will still be
runnable as `jiri-foo` instead of `jiri foo`, and I imagine that if
anyone is using this feature, it's only a few power users who are
capable of figuring out how to fix themselves.

Change-Id: Ib30015c76ba6f021fa10411a2c1c188efe23322a
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1089653
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Danielle Kay <danikay@google.com>