[run_hooks] Fetch packages before running hooks.
A hook was added which relied on the CIPD packages being present in the
checkout, which would determine what tests were generated for the ctf
builder. This resulted in clean builds like go/bbid/8740215107966775345
not generating the full set of tests that it should have as compared to
go/bbid/8740216057643304657 which had a cache hit and thus already had
a version of the CIPD packages available.

Change-Id: I26ed2bedbbecf2dc401e60addfccf8cc2b78f54b
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1097151
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>