[subcommands] Fix exit code propagation
Previously, when trying to determine whether an error was an
`ErrExitCode` that specified an exit code to propagate, this logic
checked if `err` was a `*cmdline.ErrExitCode`. However, that type's
`Error()` function takes a non-pointer receiver, so
`*cmdline.ErrExitCode` is not actually a valid error type and is not
used anywhere; instead, `cmdline.ErrExitCode` is the correct error type.

The result was that errors that were supposed to cause Jiri to exit with
specific exit codes were instead treated as generic errors, always
causing it to produce an exit code of 1, which in turn broke some
infrastructure logic that checks Jiri return codes in order to produce
better error messages about failures (e.g.
https://ci.chromium.org/b/8731153877381176257).

Fixes a bug introduced in https://fxrev.dev/1095993.

Change-Id: Ie5a4bd8770feef9ffb2575757132897a26cd2e2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1159032
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Danielle Kay <danikay@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>