[jiri] Skip packages without expansions on non-matching platforms
Our toolchain allows for cross compiling Fuchsia host tools to another
architecture, such as using a fast linux-amd64 device to compile
artifacts for a linux-arm64, so we need to be able to download the
supporting packages that don't match the current platform. This is done
by explicitly specifying the full package name with something like:

```
    <package name="fuchsia/third_party/llvm/linux-arm64"
             version="git_revision:..."
             platforms="linux-amd64,linux-arm64"
             path="prebuilt/third_party/llvm/linux-arm64"/>

    <package name="fuchsia/third_party/llvm/mac-amd64"
             version="git_revision:..."
             platforms="mac-amd64"
             path="prebuilt/third_party/llvm/mac-x64"/>
```

However, we don't support cross compiling from linux to mac, but before
this patch we would always download the mac artifacts on linux because
`cipd` will only filter out packages if it uses a ${...} substitution
that doesn't match the current system.

This patch avoids this situation by checking if the package has a
platform constraint with no substitutions, and just generates a blank
line for the cipd.ensure file.

As of writing, this patch avoids downloading about 1.5 GiB of artifacts.

Fixed: 408079132
Change-Id: I06557b86a52a57e9793b9e9f05a45bdc65a90b08
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1255264
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>