[jiri] Only print analytics log if analytics enabled
This log was introduced in
commit 5f91dda2e2d659761c803ea6f240024fa4110a5b.

The immediate motivation for this is to fix `jiri manifest` calls made
by infra, which expect the stdout to be valid JSON. Since infra runs
Jiri in verbose (-vv) mode, stdout includes trace logging, which
corrupts the JSON output.

This happened to be the only log emitted by `jiri manifest`, and infra
doesn't enable Jiri analytics, so making it conditional on analytics
being enabled should be sufficient to avoid it being printed in infra.

A better long-term fix would be to add a flag to `jiri manifest` to
specify a file to which infra should write the output, but that would
require some cross-repo changes, so going with the easy fix for now to
unblock Jiri rolls into infra.

Change-Id: I9ff1763c25731e1b5b5ff8e89fdea61b94e98c90
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1269004
Reviewed-by: Marc Khouri <mnck@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>