Remove tracking from branches created by `jiri patch`
Previously, when `jiri patch` created a branch for the CL under test, it
would set up the branch to track the project's `remote_branch` (default
"main").

A subsequent `jiri update` would then try to rebase (if
`-rebase-tracked` was set) or fast-forward merge the branch in order to
include any unrolled commits on the tracking branch.

However, this causes issues for the new atomic commits workflow, where
we don't want to fast-forward a project from JIRI_HEAD to the tip of its
main branch unless we do that for all projects, since unrolled commits
may depend on unrolled commits in other projects.

Rather than change the behavior of `jiri update` to disable the
fast-forwarding, which is risky because `jiri update` is used by humans,
just change the behavior of `jiri patch`, which is primarily used by
infra and thus has a more narrow set of use cases that's easier to
validate.

Bug: 426620623
Change-Id: Id88351a1f5b232a6fc008883dd05f6f729d13e8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1306924
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>