[project] Support moving projects into subdirectories
Suppose a project located in "third_party/foo" contained a directory
called "src". In this case, changing the project's path from
"third_party/foo" to "third_party/foo/src" would fail because the
destination directory already existed.

However, this move is actually safe - we're just moving the entire
directory down a level. So add a check to handle this special case.

Also add an integration test to make sure it works as intended. To get
the test to reproduce the failure, I needed to update the code that
calls `op.Test()` to actually call `Test()` on all operations in the
batch - previously it was skipping the first operation in the batch, and
so the test wasn't able to reproduce the failure because the test case
produced a batch of size one.

Change-Id: I557bdb436205ae128a2f0976d1f965e4b41c00b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1348544
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>