[retry] Fix backoff calculation logic.
Previously, the backoff timeout was calculated using incorrect units.
`e.InitialInterval` was in time.Duration (nanoseconds), so `next`
was in nanoseconds, but it was compared with `e.MaxInterval` which
was a value in seconds.

As a result, in `retry.Function()` we always had to wait for 64 seconds
even if it's the first retry.

This change fixes the backoff calculation logic and makes all duration
variables use the `time.Duration` type.

Test: go test go.fuchsia.dev/jiri/retry
Test: `jiri update` waits for 5 seconds on its first git failure.
Bug: 405410289
Change-Id: I4267bfcec4a9cca3fce459f7a44c28170a40d0cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1386537
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>