[submodules] Drop support for submodules entirely
We disabled submodules in fuchsia checkouts in January 2025 (10 months
ago). Submodule support in Jiri was only kept around as on offramp so
that people's checkouts would automatically be migrated off submodules
the next time they ran `jiri update`.

As long as a developer has run `jiri update` in the last 10 months
(which should be the case for the vast majority of Fuchsia engineers)
their checkout will no longer be on submodules.

If a developer hasn't run `jiri update` in the last 10 months, they'll
need to manually use an older version of Jiri (prior to this change) to
roll back submodules, or delete and recreate their checkout.

Bug: 386810791
Change-Id: I5be6bab9b2868502be7825308729d91e4c138401
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/1429434
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>